fix: resolve PHPStan type safety errors across 7 files
- Cast avg('rating') to float in ReviewsController
- Add ?? 0 default for nullable quoted_amount in catering quote email
- Add ?? 0 for nullable products_count in Str::plural()
- Add ?? '' for nullable comment in Str::limit() (Hero, Reviews components)
- Fix FinancialCalculator monthlyBreakdown return type annotation
- Fix DescriptionGeneratorService str_replace return type

+++ app/Http/Controllers/Storefront/ReviewsController.php
@@ -19,7 +19,7 @@ class ReviewsController extends Controller
             ->latest()
             ->paginate(12);
 
-        $avgRating = Review::query()->approved()->avg('rating');
+        $avgRating = (float) Review::query()->approved()->avg('rating');
         $totalReviews = Review::query()->approved()->count();
 
         $storeName = settings('store_name', 'Our Bakery');
